Fix recognition of new rows
When new rows are added, they are not recognized by the spreadsheet, bringing bad behaviors like: if we have a formula with a range and we add a new intermediate row in this range, the new row is not recognized by the formula; if we have a column that is hidden by hideColumn() and we add a row and then set a value to this hidden column, the content is showed on visible column by the side like innerHTML. After call setData, the spreadsheet works nice. So, I removed the part of initialize the history container and controller, cleaning the setData from this scope, creating a new function initializeHistory and distribuiting together with setData over the code, to keep funcionality and place the setData at the ending of updateTableReferences(), undo() and redo().